% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba.R
\name{.AddHTTPAttributes}
\alias{.AddHTTPAttributes}
\title{Adds additional data from the HTTP response to the dataframe.}
\usage{
.AddHTTPAttributes(dframe, tba_response)
}
\arguments{
\item{dframe}{A dataframe.}

\item{tba_response}{A list returned by `.SendRequest()`.}
}
\value{
A dataframe with additional attributes.
}
\description{
Adds additional data from the HTTP response to the dataframe.
}
