% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{.CallTBA}
\alias{.CallTBA}
\title{Submits an HTTP request to the TBA API.}
\usage{
.CallTBA(arg_list, api_calls, extent = "full")
}
\arguments{
\item{arg_list}{A list of TBA arguments.}

\item{api_calls}{A list of API calls.}

\item{extent}{Either "full", "simple", or "keys". Specifies the number
of columns that will be returned.}
}
\value{
The JSON text returned from the TBA API.
}
\description{
Submits an HTTP request to the TBA API.
}
