% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{.ClassifyArg}
\alias{.ClassifyArg}
\title{Identifies a single argument's parameter type}
\usage{
.ClassifyArg(arg)
}
\arguments{
\item{arg}{A length-one character vector}
}
\value{
A named character vector, where the value is the original
  argument and the name is the parameter.
}
\description{
Takes any user-supplied argument and identifies it's parameter. In
other words, identifies if an argument is a year, team key, event key,
match key, etc.
}
