% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{.ClassifyArgs}
\alias{.ClassifyArgs}
\title{Identifies Arguments' Parameter Type}
\usage{
.ClassifyArgs(arg_list)
}
\arguments{
\item{arg_list}{A character vector containing argument values.}
}
\value{
A named character vector, where the values are the original
  argument and the names are the parameter. The return value can
  be passed to `.BuildApiCommand()`'s *args* parameter.
}
\description{
Takes a list of user-supplied arguments and identifies their
parameters. In other words, identifies if an argument is a year,
team key, event key, match key, etc. This function uses
`.ClassifyArg()` to convert each element in the list.
}
