% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba.R
\name{.ConvertUnixTsColumn}
\alias{.ConvertUnixTsColumn}
\title{Converts column of Unix time stamps to readable date-time objects.}
\usage{
.ConvertUnixTsColumn(col)
}
\arguments{
\item{col}{A dataframe column of Unix time stamps.}
}
\value{
A dataframe column of POSIXct date-time objects.
}
\description{
TBA formats date-times as Unix time stamps, which is the elapsed
seconds since 12:00 AM, January 1st, 1970. The date times are in
coordinated universal time. Unless your competition is being held
in Greenland, Iceland, or West Africa, you'll need to convert the
date times to local time.
}
