% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{.SendRequest}
\alias{.SendRequest}
\title{Sends an HTTP Request}
\usage{
.SendRequest(url_path, key_service = "tba_auth_key")
}
\arguments{
\item{url_path}{Text that is added to the base TBA URL that specifies
the TBA command.}

\item{key_service}{The name of the key, as stored by the keyring package.}
}
\value{
A named list with the following elements:
  * text: The content of the HTTP response.
  * URL: The URL to which the HTTP request was sent.
  * date: The date and time when the data was retrieved from TBA.
  * last_modified: The content of the HTTP response's last-modified
    header, showing when the TBA data was last updated.
}
\description{
Sends an HTTP request to The Blue Alliance (TBA) Read API V3. The TBA
authorization key must be stored by the *keyring* package. If not
specified as an argument, assumes the *keyring* service name is
"tba_auth_key".
}
\details{
.SendRequest is called by other functions in the TbaR package.

.SendRequest raises an http error if the HTTP request is unsuccessful.
}
