% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba.R
\name{.UnnestColumn}
\alias{.UnnestColumn}
\title{Converts a column containing a list to several dataframe columns.}
\usage{
.UnnestColumn(dframe, col_name)
}
\arguments{
\item{dframe}{A datframe}

\item{col_name}{The name of the dataframe column that will be un-nested.}
}
\value{
A dataframe
}
\description{
`.UnnestColumn()` checks if the dataframe contains the column to be
un-nested. If not, the dataframe is returned unaltered.
}
