% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba.R
\name{GetDistricts}
\alias{GetDistricts}
\title{Retrieves a dataframe of FIRST competition districts}
\usage{
GetDistricts(year_or_teamkey)
}
\arguments{
\item{year_or_teamkey}{A four-digit year or team key, i.e., 2019 or
"frc1318". The year can be an integer or string.}
}
\value{
A dataframe.
}
\description{
If a year is provided, returns all districts that were in effect
during that year. If a team key is provided, returns all districts
that have had that team as a member.
}
\examples{
GetDistricts(2019)
GetDistricts("frc2976")
}
