% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba.R
\name{GetMatch}
\alias{GetMatch}
\title{Gets data for a single FRC match.}
\usage{
GetMatch(match_key, extent = "full")
}
\arguments{
\item{match_key}{A match key, e.g., 2020wasno_qm_42}

\item{extent}{Either "full" or "simple".}
}
\value{
A named list.
}
\description{
Gets data for a single FRC match.
}
\examples{
GetMatch("2020waspo_qf2m1")
}
