% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba.R
\name{GetStatus}
\alias{GetStatus}
\title{Retrieves the status of the TBA Read API (V3)}
\usage{
GetStatus()
}
\value{
A named list.
}
\description{
Retrieves the status of the TBA Read API (V3)
}
\examples{
GetStatus()
}
